# Speed up annotation retrieval (for pre-populating empty manifests)