--- a/.gitignore
+++ b/.gitignore
@@ -115,6 +115,7 @@ celerybeat.pid
 # Environments
 .env
 .venv
+.envrc
 env/
 venv/
 ENV/