--- a/.gitignore
+++ b/.gitignore
@@ -150,6 +150,7 @@ dmypy.json
 
 # Google services authorization credentials file
 credentials.json
+schematic_service_account_creds.json
 
 # token.pickle created after authorization
 token.pickle
@@ -159,3 +160,6 @@ clean.sh
 
 # direnv file
 .envrc
+
+# Intermediate files
+data/json_schema_logs/json_schema_log.json