I added support for using file views to batch-retrieve annotations for large datasets, which I've tentatively defined as datasets containing over 50 files/folders. Compared to my benchmarking in #411, the code is much faster. Here's an updated [profile.txt](https://github.com/Sage-Bionetworks/schematic/files/6078989/profile.txt). For context, the new "batch mode" is implemented in `getDatasetAnnotationsBatch()` and is called from within `getDatasetAnnotations()`. 

The `getDatasetAnnotations()` function only takes 63 seconds now (instead of 527 seconds using the original approach, or almost 9 minutes). In fact, the annotation retrieval isn't the bottleneck anymore. The `getDatasetAnnotationsBatch()` call only took 15 seconds. From what I can tell, most of the time in `getDatasetAnnotations()` was spent on running `getFilesInStorageDataset()`, which takes 45 seconds on average. This could be cut down significantly by relying on the file view itself to build a list of dataset files, but that'll have to wait for a separate PR. 

Another easy point of improvement is the `getDatasetManifest()` function, which also calls `getFilesInStorageDataset()`. I believe @milen-sage has a more efficient version of `getDatasetManifest()` laying around, so once that's reviewed and merged, that should speed things up by roughly ~40% (once this PR is merged). 

(🌺 Edited by @allaway: Closes #424)