@milen-sage 2021-03-03
Yes, thanks @BrunoGrandePhD for the reminder - I'll commit the more efficient getDatasetManifest soon if not today - tomorrow.