@BrunoGrandePhD 2021-03-03
The tests all pass locally. I think the failures on GitHub are due to concurrency because each worker is creating/getting the same file view and one of them deletes it before the others can query it. I can readily resolve this by including a random string in the file view name. This works for now since these file views aren't intended to be user-facing and are deleted shortly after being created. 

Not sure if anyone has a better solution to prevent this kind of clash. 

![image](https://user-images.githubusercontent.com/740725/109871736-54e4a380-7c20-11eb-8966-57e2f62147f5.png)