@BrunoGrandePhD 2021-03-24
Remaining tasks for @BrunoGrandePhD: 

- [ ] Run `cProfile` benchmark using minimal data model and NF's large dataset
- [ ] Resolve pending merge conflicts

Then, this PR can be merged! 🚀