@milen-sage 2021-04-07
Thank you for the detailed notes and benchmarking @BrunoGrandePhD !

There is daily and weekly variability in Google's API response time (independent of batch vs. individual requests); however reading through the timing you posted above ~30s to get an empty manifest is a lot: we might want to doublecheck if we are not doing something wrong. Agreed that requires more profiling and a separate PR.

It looks like you can check the box [here](https://github.com/Sage-Bionetworks/schematic/pull/419#issuecomment-806165585) and we can merge during/after today's CR :) ?