@sujaypatil96 2021-04-07
Thank you for the detailed testing and benchmarking on the speed up @BrunoGrandePhD. I ran the above command on the NF data model as well, and got [similar benchmarking results](https://github.com/Sage-Bionetworks/schematic/files/6274030/profile-1.txt).

Overall runtime was ~106 s.