@allaway 2021-04-08
🚀 

I checked out this PR and it fixes the issue I was encountering before. As you note, the `Filename` is not populating in this branch but hopefully between this and #425, we'll be able to update our production version in NF. Thanks so much! (insert party parrots and TYSM slackmojis)