@allaway 2021-04-08
Also, this branch is blazing fast in manifest generation :)